bringup of install script
Basically what it says. This is an install script that can get called via shell (and we can curl it!).
A few niceties:

* Error handling for various problems
* Checks if other frameworks are installed, and backs them up
* Changes shell to zsh, without sudo

Its the oh-my-zsh installer, but very heavily modified. I wrote my own installer at first, but decided that reinventing the wheel was stupid.

There is one modification: Zim's template files are no longer prepended to .zshrc and .zlogin, but instead, .zshrc and .zlogin are backed up. During the writing and testing of the script I've had the prepending command mess up and paste the template lines in the middle of other lines in my .zshrc, multiple times. You simply don't want to risk that.