[git] Remove `glb` alias and add `glO`
as there's actually no "brief" format originally in git (by "originally"
I mean the ones listed for `--format=<format>` in the `git log` help).
Also, we already fit all the information that `glb` had in the `glG`.
So add `glO`, that is a one-line variant of `glG` (which displays in
graph format). Now we get nice symmetrical formats in `glg`/`glG` and
`glo`/`glO`.