Add Ruby Module.
Ruby is a popular language, and has a number of tools in its ecosystem
which support it.

RVM
-----

The most popular tool is RVM which lets you manage a multiplicity of
ruby interperters on your system in the event that you have a bunch of
different projects each requiring their own interpreter. If you have
RVM installed, this module sets it up for you.

Bundler
------

Lots of handly aliases for working with bundler which, is basically a
requirement for doing any sort of Ruby development.