Add reinstall action
that removes and then installs again the modules that failed any of the
"pre" checks. This runs the "pre" checks (as "prereinstall") and, if the
checks failed (return code is not 0), then proceeds with the reinstall
for each module.

Also minor change of the use of the word "upgrade", reserving it to mean
upgrading zimfw.

Fixes #542